Add bottomTab
