+++ DarwinDemo/ActivityViews/ActivityView.swift
@@ -1,0 +1,23 @@
+//
+//  ActivityView.swift
+//  DarwinDemo
+//
+//  Created by Shota Iwamoto on 2024-11-04.
+//
+
+import SwiftUI
+
+struct ActivityView: View {
+    var body: some View {
+        ZStack{
+            Color.yellow.edgesIgnoringSafeArea(.all)
+            Text("ActivityView").foregroundColor(.white)
+        }
+    }
+    
+}
+
+#Preview {
+    ActivityView()
+}
+
