+++ DarwinDemo/DiscoveryViews/DiscoveryView.swift
@@ -1,0 +1,23 @@
+//
+//  DiscoveryView.swift
+//  DarwinDemo
+//
+//  Created by Shota Iwamoto on 2024-11-04.
+//
+
+import SwiftUI
+
+struct DiscoveryView: View {
+    var body: some View {
+        ZStack{
+            Color.blue.opacity(1.3).edgesIgnoringSafeArea(.all)
+            Text("DiscoratoryView").foregroundColor(.white)
+        }
+    }
+    
+}
+
+#Preview {
+    DiscoveryView()
+}
+
