+++ DarwinDemo/ReportsViews/ReportsView.swift
@@ -1,0 +1,22 @@
+//  ReportsView.swift
+//  DarwinDemo
+//
+//  Created by Shota Iwamoto on 2024-11-04.
+//
+
+import SwiftUI
+
+struct ReportsView: View {
+    var body: some View {
+        ZStack{
+            Color.gray.edgesIgnoringSafeArea(.all)
+            Text("ReportsView").foregroundColor(.white)
+        }
+    }
+    
+}
+
+#Preview {
+    ReportsView()
+}
+
